Add a GitHub Actions workflow to run tests
This workflow runs the unit tests for the project and uploads a
code coverage report to Codecov.

The workflow is run for every push to the `master` branch as well as
every time a pull request is opened for the `master` branch.

The `build` job is run against Node.js versions 8.x through to 15.x.